reintegrate extra env for unit tests
